fix: model 모듈로 분리
