+++ model/chatbot_model.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+class ChatbotRequest(BaseModel):
+    user_id: str
+    message: str
+
+class ScheduleResponse(BaseModel):
+    user_id: str
+    schedule: str
